Missing database.sql fix

+++ scripts/remote-drush-commands_deploy.sh
@@ -3,6 +3,15 @@
 # We run our drush commands here instead of in the .gitlab-ci.yml file because
 # the default version of php on the server doesn't always match the version we need.
 
+function update_database() {
+  cp ../dumps/$2.database.sql ../database.sql
+  drush sql-drop -y
+  drush sql-cli < ../database.sql
+  drush state:set system.maintenance_mode 1 --input-format=integer
+  drush updb -y
+  drush cr
+}
+
 function main () {
   # Make sure there is a bin dir in the home directory.
   test -d $HOME/bin || mkdir $HOME/bin
@@ -27,16 +36,15 @@ function main () {
   # Update the database.
   cd $HOME/www/$1/web
   drush state:set system.maintenance_mode 1 --input-format=integer
-  db_md5_old=($(md5sum "../database.sql"))
-  db_md5_new=($(md5sum "../dumps/$2.database.sql"))
-  if [[ "$db_md5_old" != "$db_md5_new" ]]; then
-    mv ../database.sql ../db_backups/database_$(date +%Y-%m-%d).sql
-    cp ../dumps/$2.database.sql ../database.sql
-    drush sql-drop -y
-    drush sql-cli < ../database.sql
-    drush state:set system.maintenance_mode 1 --input-format=integer
-    drush updb -y
-    drush cr
+  if [ ! -e ../database.sql ]; then
+    update_database "$@"
+  else
+    db_md5_old=($(md5sum "../database.sql"))
+    db_md5_new=($(md5sum "../dumps/$2.database.sql"))
+    if [[ "$db_md5_old" != "$db_md5_new" ]]; then
+      mv ../database.sql ../db_backups/database_$(date +%Y-%m-%d).sql
+      update_database "$@"
+    fi
   fi
   drush state:set system.maintenance_mode 0 --input-format=integer
 }
